Fix an invalid cast operation in the WITHOUT ROWID parsing logic.
FossilOrigin-Name: 845153be36d5ca0ca9d930c56418fb080d906740

+++ src/build.c
@@ -967,7 +967,7 @@ void sqlite3StartTable(
     }else
 #endif
     {
-      pParse->addrCrTab = (u16)sqlite3VdbeAddOp2(v, OP_CreateTable, iDb, reg2);
+      pParse->addrCrTab = sqlite3VdbeAddOp2(v, OP_CreateTable, iDb, reg2);
     }
     sqlite3OpenMasterTable(pParse, iDb);
     sqlite3VdbeAddOp2(v, OP_NewRowid, 0, reg1);
